refactor: tighten web console dreaming review structure
Extract dream action presentation, surface proposal staging audits on
detail, and settle mixed-key flags atomically on the server.
