+++ src/adapters/db/dreaming-port.ts
@@ -16,6 +16,7 @@ import {
   getActiveProfileSnapshot,
   getDailyDreamCost,
   getDreamProposal,
+  getDreamProposalStagingActionDetails,
   getLastDreamRun,
   getDreamRunActions,
   getDreamRunProposals,
@@ -61,6 +62,7 @@ export function createDreamPort(executor: SqlExecutor): DreamPort {
     getRunActions: async (runId) => getDreamRunActions(executor, runId),
     getRunProposals: async (runId) => getDreamRunProposals(executor, runId),
     getProposal: async (proposalId) => getDreamProposal(executor, proposalId),
+    getProposalStagingActionDetails: async (proposalId) => getDreamProposalStagingActionDetails(executor, proposalId),
     reviewProposal: async (input) => reviewDreamProposal(executor, input),
     listProposalBacklog: async (query) => listDreamProposalBacklog(executor, query),
     getHealthStats: async (now) => getDreamHealthStats(executor, now),
