+++ web/src/pages/DreamingPage.tsx
@@ -2,25 +2,22 @@ import { useEffect, useRef, useState } from "react";
 import { useNavigate } from "react-router-dom";
 
 import { ApiError, api } from "../api/client";
-import type { DreamJobSnapshot, DreamProposal, DreamRunActionView, DreamRunRecord, DreamRunsResponse } from "../api/types";
+import type { DreamJobSnapshot, DreamProposal, DreamRunRecord, DreamRunsResponse } from "../api/types";
 import { DataTable } from "../components/DataTable";
-import { Badge, Button, Card, CardBody, CardHeader, Chip, Drawer, Field, Input, KeyValue, Spinner, StatusDot } from "../components/primitives";
+import {
+  DreamActionChangeSummary,
+  FlagReviewActionSummary,
+  formatActionReasoning,
+  formatActionTypeLabel,
+} from "../components/dreaming/ActionReviewPanels";
+import { Badge, Button, Card, CardBody, CardHeader, Drawer, Field, Input, KeyValue, Spinner, StatusDot } from "../components/primitives";
 import { ErrorCard, RequireInstance, Skeleton } from "../components/states";
 import { useToast } from "../components/Toast";
 import { useAsync } from "../hooks/useAsync";
 import { useDreamStream } from "../hooks/useDreamStream";
 import { useInstances } from "../state/InstanceContext";
-import {
-  buildAliasReviewSummary,
-  formatAliasAutoApplyBlocker,
-  formatClaimKeyList,
-  renderAliasReviewSummaryLines,
-  summarizeAliasEvidence,
-  summarizeAliasProfiles,
-  type AliasReviewSummaryView,
-} from "../lib/alias-audit";
 import { describeEvent } from "../lib/dream-progress";
-import { formatCost, formatDateTime, formatDateTimeWithRelative, formatIssueKind, formatPercent, formatRelative, titleCase } from "../lib/format";
+import { formatCost, formatDateTime, formatDateTimeWithRelative, formatIssueKind, formatRelative, titleCase } from "../lib/format";
 import { jobStatusVariant, runStatusVariant } from "../lib/status";
 
 /** Available dreaming tiers with descriptions. */
@@ -29,33 +26,6 @@ const TIERS: { id: "light" | "standard" | "deep"; label: string; hint: string }[
   { id: "standard", label: "Standard", hint: "Reconcile + project profile" },
   { id: "deep", label: "Deep", hint: "Full extract and synthesis" },
 ];
-
-/** UUID text emitted by storage-level audit payloads. */
-const UUID_PATTERN = /^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$/iu;
-
-/** UUID fragments embedded inside storage-level audit prose. */
-const UUID_IN_TEXT_PATTERN = /\b[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}\b/giu;
-
-/** Detail field labels for operator-facing dream action payloads. */
-const DETAIL_KEY_LABELS: Record<string, string> = {
-  claim_key: "Claim key",
-  evidence_refs: "Evidence",
-  predecessor_id: "Previous durable",
-  successor_id: "Replacement durable",
-  valid_to: "Valid until",
-  alias_entity_prefix: "Alias entity",
-  alias_current_claim_keys: "Current keys",
-  alias_proposed_claim_key: "Proposed key",
-  alias_deterministic_confidence: "Deterministic confidence",
-  alias_deterministic_auto_apply_eligible: "Deterministic auto-apply",
-  alias_unresolved_reason: "Blocker",
-  alias_llm_same_slot: "LLM same slot",
-  alias_llm_canonical_claim_key: "LLM canonical key",
-  alias_llm_confidence: "LLM confidence",
-  alias_llm_rationale: "LLM rationale",
-  alias_evidence: "Alias evidence",
-  alias_key_profiles: "Alias profiles",
-};
 
 /**
  * Dreaming Runs page: launch maintenance runs and monitor live progress.
@@ -399,412 +369,3 @@ function RunDetailDrawer({ run, onClose }: { run: DreamRunRecord; onClose: () =>
     </Drawer>
   );
 }
-
-/** Plain-language summary for a proposal staged for human review. */
-function FlagReviewActionSummary({ action, onOpenProposals }: { action: DreamRunActionView; onOpenProposals: () => void }): React.ReactElement {
-  const details = action.details ?? {};
-  const currentClaimKeys = readStringArrayDetail(details.current_claim_keys);
-  const proposedClaimKeys = readStringArrayDetail(details.proposed_claim_keys);
-  const confidence = typeof details.confidence === "number" ? details.confidence : null;
-  const eligibleForApply = details.eligible_for_apply === true;
-  const blocker = typeof details.auto_apply_blocker === "string" ? details.auto_apply_blocker : null;
-  const durableById = new Map(action.durables.map((durable) => [durable.id, durable]));
-  const aliasSummary = buildAliasReviewSummary(details);
-
-  return (
-    <div className="review-summary">
-      <div className="review-summary__plain">
-        <strong>
-          {aliasSummary ? "Dreaming found a possible claim-key alias cluster." : "Dreaming found a suggested claim-key change."}
-        </strong>
-        <span>
-          {aliasSummary
-            ? "Review means deciding whether the current keys represent the same durable slot. Apply writes the proposed key after a backup; reject leaves the keys unchanged and records the reason."
-            : "Review means deciding whether the affected memory belongs under the proposed key. Apply writes that key after a backup; reject leaves the memory unchanged and records the reason."}
-        </span>
-      </div>
-
-      <div className="review-summary__grid">
-        <ReviewSummaryRow label={aliasSummary ? "Current keys" : "Current key"} value={formatClaimKeyList(currentClaimKeys, "(no current key)")} />
-        <ReviewSummaryRow label="Proposed key" value={formatClaimKeyList(proposedClaimKeys, "(no proposed key)")} emphasized />
-        {confidence !== null ? <ReviewSummaryRow label="Confidence" value={formatPercent(confidence)} /> : null}
-        <ReviewSummaryRow label="Decision" value={formatReviewDecisionText(eligibleForApply, blocker)} />
-      </div>
-
-      {aliasSummary ? <AliasReviewSummary summary={aliasSummary} /> : null}
-
-      <div className="stack" style={{ gap: "var(--space-2)" }}>
-        <span className="section-title">What to check</span>
-        {aliasSummary ? (
-          <ul className="review-checklist">
-            <li>The current keys describe the same stable slot for the same entity.</li>
-            <li>The proposed key is the best canonical target for all affected durables.</li>
-            <li>Similar wording alone is not enough if the facts belong to different attributes.</li>
-          </ul>
-        ) : (
-          <ul className="review-checklist">
-            <li>The memory text is really about the proposed topic.</li>
-            <li>The proposed key describes the stable slot this memory should be grouped under.</li>
-            <li>Applying the key would group it with the right related memories, not just similar wording.</li>
-          </ul>
-        )}
-      </div>
-
-      <ReviewEvidenceSummary details={details} />
-
-      {action.durableIds.length > 0 ? (
-        <div className="row wrap" style={{ gap: "var(--space-2)" }}>
-          <span className="muted" style={{ fontSize: "var(--text-xs)" }}>Affected</span>
-          {action.durableIds.map((durableId) => (
-            <Chip key={durableId} className="chip--compact" title={formatAffectedDurableReferenceTitle(durableId, action.details, durableById)}>
-              <span className="truncate">{formatAffectedDurableReference(durableId, action.details, durableById)}</span>
-            </Chip>
-          ))}
-        </div>
-      ) : null}
-
-      {action.durables.length > 0 ? <ActionDurableList action={action} /> : null}
-
-      {hasActionDetails(action) ? (
-        <details className="diagnostics review-summary__diagnostics">
-          <summary>Diagnostics</summary>
-          <ActionDetailsGrid action={action} />
-        </details>
-      ) : null}
-
-      <div className="review-summary__actions">
-        <Button variant="primary" size="sm" icon="arrow-right" onClick={onOpenProposals}>
-          Open review
-        </Button>
-      </div>
-    </div>
-  );
-}
-
-/** One row in the staged-review summary. */
-function ReviewSummaryRow({ label, value, emphasized = false }: { label: string; value: string; emphasized?: boolean }): React.ReactElement {
-  return (
-    <div className="review-summary__row">
-      <span>{label}</span>
-      <strong className={emphasized ? "review-summary__value--emphasized" : undefined}>{value}</strong>
-    </div>
-  );
-}
-
-/** Renders the structured alias-convergence audit in a compact operator view. */
-function AliasReviewSummary({ summary }: { summary: AliasReviewSummaryView }): React.ReactElement {
-  const lines = renderAliasReviewSummaryLines({ summary, formatClaimKeys: formatClaimKeyList });
-  return (
-    <div className="review-evidence">
-      <span className="section-title">Alias convergence audit</span>
-      {lines.map((line) => (line ? <span key={line.key}>{line.text}</span> : null))}
-    </div>
-  );
-}
-
-/** Plain-language evidence summary for review proposals. */
-function ReviewEvidenceSummary({ details }: { details: Record<string, unknown> }): React.ReactElement | null {
-  const familyReuseCount = readNumberDetail(details.support_family_reuse_count);
-  const groundedFamilyReuseCount = readNumberDetail(details.support_grounded_family_reuse_count);
-  const supportEvidence = readStringArrayDetail(details.support_evidence);
-  const hasSupportedCandidate = details.supported_candidate === true;
-
-  const lines = [
-    familyReuseCount > 0
-      ? `${familyReuseCount} related memor${familyReuseCount === 1 ? "y already uses" : "ies already use"} a compatible key.`
-      : null,
-    groundedFamilyReuseCount > 0
-      ? `${groundedFamilyReuseCount} of those related memories include supporting provenance.`
-      : null,
-    hasSupportedCandidate || supportEvidence.length > 0 ? "Dreaming found matching evidence in the existing corpus." : null,
-  ].filter((line): line is string => line !== null);
-
-  if (lines.length === 0) {
-    return null;
-  }
-
-  return (
-    <div className="review-evidence">
-      <span className="section-title">Why it was suggested</span>
-      {lines.map((line) => (
-        <span key={line}>{line}</span>
-      ))}
-    </div>
-  );
-}
-
-/** Renders the concrete change payload for one dream action. */
-function DreamActionChangeSummary({ action }: { action: DreamRunActionView }): React.ReactElement {
-  const durableById = new Map(action.durables.map((durable) => [durable.id, durable]));
-
-  return (
-    <div className="stack" style={{ gap: "var(--space-2)", marginTop: "var(--space-3)" }}>
-      {action.durableIds.length > 0 ? (
-        <div className="row wrap" style={{ gap: "var(--space-2)" }}>
-          <span className="muted" style={{ fontSize: "var(--text-xs)" }}>Affected</span>
-          {action.durableIds.map((durableId) => {
-            const label = formatAffectedDurableReference(durableId, action.details, durableById);
-            return (
-              <Chip key={durableId} className="chip--compact" title={formatAffectedDurableReferenceTitle(durableId, action.details, durableById)}>
-                <span className="truncate">{label}</span>
-              </Chip>
-            );
-          })}
-        </div>
-      ) : null}
-
-      {action.durables.length > 0 ? <ActionDurableList action={action} /> : null}
-
-      {hasActionDetails(action) ? (
-        <details className="diagnostics">
-          <summary>Diagnostics</summary>
-          <ActionDetailsGrid action={action} />
-        </details>
-      ) : null}
-    </div>
-  );
-}
-
-/** Renders action detail fields as an audit grid. */
-function ActionDetailsGrid({ action }: { action: DreamRunActionView }): React.ReactElement | null {
-  const details = action.details ? Object.entries(action.details).filter(([, value]) => value !== null && value !== undefined && value !== "") : [];
-  const durableById = new Map(action.durables.map((durable) => [durable.id, durable]));
-
-  if (details.length === 0) {
-    return null;
-  }
-
-  return (
-    <div className="change-grid">
-      {details.map(([key, value]) => {
-        const formatted = formatDetailValue(value, key, durableById);
-        return (
-          <div key={key} style={{ display: "contents" }}>
-            <span className="change-grid__key">{formatDetailKey(key)}</span>
-            <span className="change-grid__value truncate" title={formatted.title}>
-              {formatted.label}
-            </span>
-          </div>
-        );
-      })}
-    </div>
-  );
-}
-
-/** Returns whether an action has raw detail fields worth exposing in diagnostics. */
-function hasActionDetails(action: DreamRunActionView): boolean {
-  return action.details ? Object.values(action.details).some((value) => value !== null && value !== undefined && value !== "") : false;
-}
-
-/** Renders affected durable snippets for an action. */
-function ActionDurableList({ action }: { action: DreamRunActionView }): React.ReactElement {
-  return (
-    <div className="stack" style={{ gap: "var(--space-2)" }}>
-      {action.durables.map((durable) => (
-        <div key={durable.id} className="change-item">
-          <div className="spread" style={{ gap: "var(--space-2)" }}>
-            <strong className="truncate" style={{ fontSize: "var(--text-xs)" }} title={durable.subject}>{durable.subject}</strong>
-            {durable.claim_key ? (
-              <span className="mono muted truncate" style={{ fontSize: "var(--text-2xs)" }} title={durable.claim_key}>
-                {durable.claim_key}
-              </span>
-            ) : null}
-          </div>
-          <span className="muted truncate" style={{ fontSize: "var(--text-xs)" }} title={durable.content}>{durable.content}</span>
-        </div>
-      ))}
-    </div>
-  );
-}
-
-/** Formats a dreaming action type for operator display. */
-function formatActionTypeLabel(actionType: string): string {
-  return actionType === "flag_review" ? "Needs review" : titleCase(actionType);
-}
-
-/** Formats an action detail key for compact display. */
-function formatDetailKey(key: string): string {
-  return DETAIL_KEY_LABELS[key] ?? titleCase(key.replaceAll("_", " "));
-}
-
-/** Operator-facing detail value with the raw value preserved for hover text. */
-interface FormattedDetailValue {
-  label: string;
-  title: string;
-}
-
-/** Formats a durable reference as a stable operator-facing label. */
-function formatDurableReference(durableId: string, durableById: Map<string, DreamRunActionView["durables"][number]>): string {
-  const durable = durableById.get(durableId);
-  if (!durable) {
-    return "Missing durable";
-  }
-  if (durable.claim_key) {
-    return `${durable.subject} - ${durable.claim_key}`;
-  }
-  return durable.subject;
-}
-
-/** Formats an affected durable list item without exposing storage IDs. */
-function formatAffectedDurableReference(
-  durableId: string,
-  details: Record<string, unknown> | null,
-  durableById: Map<string, DreamRunActionView["durables"][number]>,
-): string {
-  const durable = durableById.get(durableId);
-  if (durable) {
-    return formatDurableReference(durableId, durableById);
-  }
-  return formatMissingDurableLabel(durableId, details);
-}
-
-/** Builds full hover text for an affected durable chip. */
-function formatDurableReferenceTitle(durableId: string, durableById: Map<string, DreamRunActionView["durables"][number]>): string {
-  const durable = durableById.get(durableId);
-  if (!durable) {
-    return `Missing durable: ${durableId}`;
-  }
-  const parts = [`${durable.subject}`, durable.claim_key ? `Claim key: ${durable.claim_key}` : null, `Durable ID: ${durable.id}`];
-  return parts.filter((part): part is string => Boolean(part)).join("\n");
-}
-
-/** Builds hover text for an affected durable list item. */
-function formatAffectedDurableReferenceTitle(
-  durableId: string,
-  details: Record<string, unknown> | null,
-  durableById: Map<string, DreamRunActionView["durables"][number]>,
-): string {
-  const durable = durableById.get(durableId);
-  if (durable) {
-    return formatDurableReferenceTitle(durableId, durableById);
-  }
-  return `${formatMissingDurableLabel(durableId, details)}: ${durableId}`;
-}
-
-/** Replaces raw durable UUIDs in persisted action prose with friendly labels. */
-function formatActionReasoning(
-  reasoning: string,
-  details: Record<string, unknown> | null,
-  durableById: Map<string, DreamRunActionView["durables"][number]>,
-): string {
-  const predecessorId = typeof details?.predecessor_id === "string" ? details.predecessor_id : null;
-  const successorId = typeof details?.successor_id === "string" ? details.successor_id : null;
-  if (predecessorId && successorId && reasoning.startsWith("Superseded durable ")) {
-    const predecessor = formatSentenceReference(formatAffectedDurableReference(predecessorId, details, durableById));
-    const successor = formatSentenceReference(formatAffectedDurableReference(successorId, details, durableById));
-    return `Superseded ${predecessor} with temporal revision ${successor}.`;
-  }
-
-  return reasoning.replace(UUID_IN_TEXT_PATTERN, (durableId) => formatAffectedDurableReference(durableId, details, durableById));
-}
-
-/** Makes role labels read naturally inside generated prose. */
-function formatSentenceReference(label: string): string {
-  if (label === "Previous durable" || label === "Replacement durable" || label === "Missing durable") {
-    return `${label.charAt(0).toLowerCase()}${label.slice(1)}`;
-  }
-  return label;
-}
-
-/** Labels missing durable references by their role when action details expose it. */
-function formatMissingDurableLabel(durableId: string, details: Record<string, unknown> | null): string {
-  if (details?.predecessor_id === durableId) {
-    return "Previous durable";
-  }
-  if (details?.successor_id === durableId) {
-    return "Replacement durable";
-  }
-  return "Missing durable";
-}
-
-/** Formats a structured action detail value for compact display. */
-function formatDetailValue(value: unknown, key: string, durableById: Map<string, DreamRunActionView["durables"][number]>): FormattedDetailValue {
-  if (Array.isArray(value)) {
-    if (key === "alias_evidence") {
-      return summarizeAliasEvidence(value);
-    }
-    if (key === "alias_key_profiles") {
-      return summarizeAliasProfiles(value);
-    }
-    const entries = value.map((entry) => formatDetailValue(entry, key, durableById));
-    return {
-      label: summarizeDetailList(entries, key),
-      title: entries.map((entry) => entry.title).join("\n"),
-    };
-  }
-
-  if (typeof value === "object" && value !== null) {
-    const raw = JSON.stringify(value);
-    return { label: raw, title: raw };
-  }
-
-  const text = String(value);
-  const durable = durableById.get(text);
-  if (durable) {
-    return { label: formatDurableReference(text, durableById), title: formatDurableReferenceTitle(text, durableById) };
-  }
-  if (isTimestampDetail(key, text)) {
-    return { label: formatDateTime(text), title: text };
-  }
-  if (isEvidenceReference(text)) {
-    return { label: "Episode evidence", title: text };
-  }
-  if (UUID_PATTERN.test(text)) {
-    return { label: friendlyIdentifierLabel(key), title: text };
-  }
-  return { label: text, title: text };
-}
-
-/** Summarizes repeated action detail values without exposing raw storage IDs. */
-function summarizeDetailList(entries: FormattedDetailValue[], key: string): string {
-  if (entries.length === 0) {
-    return "-";
-  }
-  if (key === "evidence_refs") {
-    return entries.length === 1 ? entries[0]?.label ?? "Evidence" : `${entries.length} evidence refs`;
-  }
-  return entries.map((entry) => entry.label).join(", ");
-}
-
-/** Returns whether a detail value is an ISO timestamp field. */
-function isTimestampDetail(key: string, value: string): boolean {
-  if (!/(?:_at|_to|_from)$/u.test(key)) {
-    return false;
-  }
-  return Number.isFinite(Date.parse(value));
-}
-
-/** Returns whether a raw evidence reference contains an internal storage UUID. */
-function isEvidenceReference(value: string): boolean {
-  const [prefix, id] = value.split(":", 2);
-  return prefix === "episode" && id !== undefined && UUID_PATTERN.test(id);
-}
-
-/** Builds a generic friendly label for an internal identifier detail. */
-function friendlyIdentifierLabel(key: string): string {
-  const label = DETAIL_KEY_LABELS[key] ?? titleCase(key.replaceAll("_", " "));
-  return label.toLowerCase().includes("durable") ? label : `${label} record`;
-}
-
-/** Reads a string array from an untyped action-detail payload. */
-function readStringArrayDetail(value: unknown): string[] {
-  return Array.isArray(value) ? value.filter((entry): entry is string => typeof entry === "string" && entry.trim().length > 0) : [];
-}
-
-/** Reads a number from an untyped action-detail payload. */
-function readNumberDetail(value: unknown): number {
-  return typeof value === "number" && Number.isFinite(value) ? value : 0;
-}
-
-/** Describes the review action available for a proposal. */
-function formatReviewDecisionText(eligibleForApply: boolean, blocker: string | null): string {
-  if (eligibleForApply) {
-    return "Apply is available if the proposed key is correct.";
-  }
-  if (blocker) {
-    const formattedBlocker = formatAliasAutoApplyBlocker(blocker);
-    return `Reject or inspect manually. ${formattedBlocker}${/[.!?]$/u.test(formattedBlocker) ? "" : "."}`;
-  }
-  return "Reject or inspect manually. Apply is not available for this proposal.";
-}
-
